refactor code to use one general controller defined in env_models.__init__
